+++ mathics/builtin/scipy_utils/integrators.py
@@ -1,12 +1,9 @@
 # -*- coding: utf-8 -*-
 
 from mathics.core.builtin import check_requires_list
-from mathics.core.util import IS_PYPY
 
-if IS_PYPY or not check_requires_list(["scipy", "numpy"]):
+if not check_requires_list(["scipy", "numpy"]):
     raise ImportError
-
-import numpy as np
 
 
 def _scipy_interface(integrator, options_map, mandatory=None, adapt_func=None):
@@ -25,7 +22,8 @@ def _scipy_interface(integrator, options_map, mandatory=None, adapt_func=None):
                 if native_opt[1]:
                     val = native_opt[1](val)
                 native_opts[native_opt[0]] = val
-        return adapt_func(integrator(fun, a, b, **native_opts))
+        if adapt_func is not None:
+            return adapt_func(integrator(fun, a, b, **native_opts))
 
     def _scipy_proxy_func(fun, a, b, **opts):
         native_opts = {}
@@ -43,7 +41,7 @@ def _scipy_interface(integrator, options_map, mandatory=None, adapt_func=None):
 
 
 try:
-    from scipy.integrate import nquad, quad, romberg
+    from scipy.integrate import nquad, quad
 except Exception:
     scipy_nintegrate_methods = {}
 else:
@@ -66,17 +64,6 @@ else:
                     },
                     {"full_output": 1},
                     lambda res: (res[0], res[1]),
-                ),
-                False,
-            )
-        ),
-        "Romberg": tuple(
-            (
-                _scipy_interface(
-                    romberg,
-                    {"tol": ("tol", None), "maxrec": ("divmax", None)},
-                    None,
-                    lambda x: (x, np.nan),
                 ),
                 False,
             )